sct: move nullification into SCT component
Since any kind of state fragment can be nullified, this should be the
responsibility of the SCT component, not the shielded pool.
