+++ crates/bin/pcli/src/command/query/shielded_pool.rs
@@ -42,14 +42,15 @@ impl ShieldedPool {
     pub fn key(&self) -> String {
         use penumbra_compact_block::state_key as cb_state_key;
         use penumbra_sct::state_key as sct_state_key;
-        use penumbra_shielded_pool::state_key;
         match self {
             ShieldedPool::Anchor { height } => sct_state_key::anchor_by_height(*height),
             ShieldedPool::BlockAnchor { height } => sct_state_key::block_anchor_by_height(*height),
             ShieldedPool::EpochAnchor { epoch } => sct_state_key::epoch_anchor_by_index(*epoch),
             ShieldedPool::CompactBlock { height } => cb_state_key::compact_block(*height),
             ShieldedPool::Commitment { commitment } => sct_state_key::note_source(commitment),
-            ShieldedPool::Nullifier { nullifier } => state_key::spent_nullifier_lookup(nullifier),
+            ShieldedPool::Nullifier { nullifier } => {
+                sct_state_key::spent_nullifier_lookup(nullifier)
+            }
         }
     }
