+++ crates/core/component/governance/src/component/view.rs
@@ -11,8 +11,8 @@ use penumbra_asset::{asset, Value, STAKING_TOKEN_DENOM};
 use penumbra_chain::component::{StateReadExt as _, StateWriteExt as _};
 use penumbra_num::Amount;
 use penumbra_proto::{StateReadProto, StateWriteProto};
-use penumbra_sct::Nullifier;
-use penumbra_shielded_pool::component::{StateReadExt as _, SupplyRead};
+use penumbra_sct::{component::StateReadExt as _, Nullifier};
+use penumbra_shielded_pool::component::SupplyRead;
 use penumbra_stake::{DelegationToken, GovernanceKey, IdentityKey};
 use penumbra_tct as tct;
 use tokio::task::JoinSet;
