+++ crates/core/component/shielded-pool/src/component/shielded_pool.rs
@@ -2,18 +2,13 @@ use std::sync::Arc;
 
 use anyhow::Result;
 use async_trait::async_trait;
-use cnidarium::StateRead;
 use cnidarium::StateWrite;
 use cnidarium_component::Component;
-use penumbra_proto::StateReadProto;
 use penumbra_sct::CommitmentSource;
-use penumbra_sct::NullificationInfo;
-use penumbra_sct::Nullifier;
 use tendermint::v0_37::abci;
 use tracing::instrument;
 
 use crate::genesis::Content as GenesisContent;
-use crate::state_key;
 
 use super::{NoteManager, SupplyWrite};
 
@@ -72,26 +67,3 @@ impl Component for ShieldedPool {
         Ok(())
     }
 }
-
-#[async_trait]
-pub trait StateReadExt: StateRead {
-    async fn check_nullifier_unspent(&self, nullifier: Nullifier) -> Result<()> {
-        if let Some(info) = self
-            .get::<NullificationInfo>(&state_key::spent_nullifier_lookup(&nullifier))
-            .await?
-        {
-            anyhow::bail!(
-                "nullifier {} was already spent in {:?}",
-                nullifier,
-                hex::encode(&info.id),
-            );
-        }
-        Ok(())
-    }
-    async fn spend_info(&self, nullifier: Nullifier) -> Result<Option<NullificationInfo>> {
-        self.get(&state_key::spent_nullifier_lookup(&nullifier))
-            .await
-    }
-}
-
-impl<T: StateRead + ?Sized> StateReadExt for T {}
