+++ crates/core/component/shielded-pool/src/state_key.rs
@@ -1,5 +1,4 @@
 use penumbra_asset::asset;
-use penumbra_sct::Nullifier;
 use std::string::String;
 
 pub fn token_supply(asset_id: &asset::Id) -> String {
@@ -14,10 +13,6 @@ pub fn denom_by_asset(asset_id: &asset::Id) -> String {
     format!("shielded_pool/assets/{asset_id}/denom")
 }
 
-pub fn spent_nullifier_lookup(nullifier: &Nullifier) -> String {
-    format!("shielded_pool/spent_nullifiers/{nullifier}")
-}
-
 // State keys used to temporarily store payloads and nullifiers to be inserted into the compact
 // block
 
@@ -27,8 +22,4 @@ pub fn pending_notes() -> &'static str {
 
 pub fn pending_rolled_up_payloads() -> &'static str {
     "shielded_pool/pending_rolled_up_payloads"
-}
-
-pub fn pending_nullifiers() -> &'static str {
-    "shielded_pool/pending_nullifiers"
 }
